Refactoring
Prompt
------

Let's do some more refactoring. Please move the business logic (TimeSheet, Period, and related calculation code) into a separate module.

Then split off other submodules.

Result
------
...
