+++ src/lib.rs
@@ -1,0 +1,8 @@
+//! # Work Time Tracker Library
+//!
+//! This library contains the core business logic and data storage
+//! functionalities for the command-line time tracking utility.
+
+pub mod logic;
+pub mod storage;
+
